Merge the `/` and `%` operators
Having one fewer operator seems simpler, and since the stack-based model
makes multiple outputs quite natural, why not use that.

Maybe this is going to be tedious to use, but I'd like to confirm that
in practice.

This commit also organizes the notes on the list of operators into a
list themselves, to make the connection between the sentence introducing
the notes and the notes more clear.