Add note on the nature of addresses
It turns out that I don't want to add a dedicated paragraph that says
more on this, which is where I planned to refer to the section on the
type system. Still, I believe that the new placement of the section
makes this note easier to understand, should the reader care to.