Prepare to support identifiers
Because we currently rely on a self-contained method from Rust's
standard library to parse integers, we can't easily detect the case of
a token that is an integer but doesn't fit into a 32-bit word. The
addition of identifiers would have required this ability.

This commit obviates the need for this, by treating integers that are
out of bounds the same as unknown identifiers. This is going to keep the
code trivial, as support for identifiers is added.