Add partial support for references
So far, an invalid reference (one for which a matching label does not
exist) results in a panic. It should result in an effect, which I'm
going to work on next.

---

This commit includes the missing test that also covers the previously
added labels.

I have decided to re-use the already existing test for `jump`, instead
of adding a new one. While the previously tested behavior still works, I
only did test it because it made sense in terms of the order of
implementation.

Giving integer arguments to `jump` is error-prone at best, made
unnecessary by the addition of labels and references, and should best be
avoided. It is merely an exposed implementation detail, and shouldn't be
mistaken for a valid language feature.