Extract example host into dedicated crate
This makes it possible to add more dependencies there, for example for
command-line parsing and other such things, that make working on the
host easier while possibly improving its interface.

Adding those dependencies to the previous unified crate would have been
undesirable, as the interpreter should be able to run in environments
where those dependencies would not be appropriate. In the future, I even
expect it to become `no_std`, which the host can't reasonably do.